Move Tweeter::Config to Tools::TwitterConfig
